Backend/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.6
MarkupSafe==3.0.2
openai==1.52.2
packaging==24.1
pipenv==2024.3.0
platformdirs==4.2.2
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.13
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.3
Werkzeug==3.0.6
zipp==3.20.2
